pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools_scm>=8", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "add-skills"
dynamic = ["version"]
description = "A tool for managing Claude Code skills"
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
authors = [
    {name = "Ludo Technologies LLC", email = "contact@ludo-tech.org"}
]
maintainers = [
    {name = "Ludo Technologies LLC", email = "contact@ludo-tech.org"}
]
keywords = [
    "claude",
    "claude-code",
    "skills",
    "cli",
    "automation"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
requires-python = ">=3.10"
dependencies = [
    "typer>=0.9.0",
    "rich>=10.0.0",
    "python-frontmatter>=1.0.0",
    "GitPython>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "build>=0.7.0",
    "twine>=3.0.0",
    "ruff>=0.8.0",
    "pyright>=1.1.0",
]

[project.scripts]
add-skills = "add_skills.__main__:main"

[project.urls]
Homepage = "https://github.com/ludo-technologies/add-skills"
Repository = "https://github.com/ludo-technologies/add-skills"
Issues = "https://github.com/ludo-technologies/add-skills/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools_scm]
# Default configuration - gets version from git tags

[tool.ruff]
target-version = "py310"
line-length = 88
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "TID",    # flake8-tidy-imports
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["add_skills"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.pyright]
pythonVersion = "3.10"
typeCheckingMode = "standard"
include = ["src"]
exclude = ["**/__pycache__", ".venv"]
